Crate reth_libmdbx

libmdbx-rs

Rust bindings for libmdbx.

Forked from an earlier Apache licenced version of the libmdbx-rs crate, before it changed licence to GPL. NOTE: Most of the repo came from lmdb-rs bindings.

Updating the libmdbx Version

To update the libmdbx version you must clone it and copy the dist/ folder in mdbx-sys/. Make sure to follow the building steps.

# clone libmmdbx to a repository outside at specific tag
git clone https://gitflic.ru/project/erthink/libmdbx.git ../libmdbx --branch v0.7.0
make -C ../libmdbx dist

# copy the `libmdbx/dist/` folder just created into `mdbx-sys/libmdbx`
rm -rf mdbx-sys/libmdbx
cp -R ../libmdbx/dist mdbx-sys/libmdbx

# add the changes to the next commit you will make
git add mdbx-sys/libmdbx

Modules

ffi

Macros

mdbx_try_optional

Structs

CommitLatency

Commit latencies info.

Cursor

A cursor for navigating the items within a database.

Database

A handle to an individual database in an environment.

DatabaseFlags

Database options.

Environment

An environment supports multiple databases, all residing in the same shared-memory map.

EnvironmentBuilder

Options for opening or creating an environment.

EnvironmentFlags

Geometry

Info

Environment information.

ObjectLength

If you don’t need the data itself, just its length.

RO

RW

Stat

Environment statistics.

Transaction

An MDBX transaction.

WriteFlags

Write options.

Enums

EnvironmentKind

Determines how data is mapped into memory

Error

An MDBX error kind.

HandleSlowReadersReturnCode

Iter

An iterator over the key/value pairs in an MDBX database.

IterDup

An iterator over the keys and duplicate values in an MDBX database.

MaxReadTransactionDuration

The maximum duration of a read transaction.

Mode

PageSize

SyncMode

MDBX sync mode

Traits

TableObject

Implement this to be able to decode data values

TransactionKind

Type Aliases

HandleSlowReadersCallback

Handle-Slow-Readers callback function to resolve database full/overflow issue due to a reader(s) which prevents the old data from being recycled.

Result

An MDBX result.