Struct reth_libmdbx::EnvironmentBuilder

pub struct EnvironmentBuilder { /* private fields */ }

Options for opening or creating an environment.

Implementations

impl EnvironmentBuilder

pub fn set_max_read_transaction_duration( &mut self, max_read_transaction_duration: MaxReadTransactionDuration ) -> &mut Self

Set the maximum time a read-only transaction can be open.

impl EnvironmentBuilder

pub fn open(&self, path: &Path) -> Result<Environment>

Open an environment.

Database files will be opened with 644 permissions.

pub fn open_with_permissions( &self, path: &Path, mode: mdbx_mode_t ) -> Result<Environment>

Open an environment with the provided UNIX permissions.

The path may not contain the null character.

pub fn set_kind(&mut self, kind: EnvironmentKind) -> &mut Self

Configures how this environment will be opened.

pub fn write_map(&mut self) -> &mut Self

Opens the environment with mdbx WRITEMAP

See also EnvironmentKind

pub fn set_flags(&mut self, flags: EnvironmentFlags) -> &mut Self

Sets the provided options in the environment.

pub fn set_max_readers(&mut self, max_readers: u64) -> &mut Self

Sets the maximum number of threads or reader slots for the environment.

This defines the number of slots in the lock table that is used to track readers in the the environment. The default is 126. Starting a read-only transaction normally ties a lock table slot to the Transaction object until it or the Environment object is destroyed.

pub fn set_max_dbs(&mut self, v: usize) -> &mut Self

Sets the maximum number of named databases for the environment.

This function is only needed if multiple databases will be used in the environment. Simpler applications that use the environment as a single unnamed database can ignore this option.

Currently a moderate number of slots are cheap but a huge number gets expensive: 7-120 words per transaction, and every Transaction::open_db() does a linear search of the opened slots.

pub fn set_rp_augment_limit(&mut self, v: u64) -> &mut Self

pub fn set_loose_limit(&mut self, v: u64) -> &mut Self

pub fn set_dp_reserve_limit(&mut self, v: u64) -> &mut Self

pub fn set_txn_dp_limit(&mut self, v: u64) -> &mut Self

pub fn set_spill_max_denominator(&mut self, v: u8) -> &mut Self

pub fn set_spill_min_denominator(&mut self, v: u8) -> &mut Self

pub fn set_geometry<R: RangeBounds<usize>>( &mut self, geometry: Geometry<R> ) -> &mut Self

Set all size-related parameters of environment, including page size and the min/max size of the memory map.

pub fn set_log_level(&mut self, log_level: MDBX_log_level_t) -> &mut Self

pub fn set_handle_slow_readers( &mut self, hsr: HandleSlowReadersCallback ) -> &mut Self

Set the Handle-Slow-Readers callback. See HandleSlowReadersCallback for more information.