Struct reth_libmdbx::Transaction

pub struct Transaction<K>
where
    K: TransactionKind,
{ /* private fields */ }

An MDBX transaction.

All database operations require a transaction.

Implementations

impl<K> Transaction<K>
where K: TransactionKind,

pub fn env(&self) -> &Environment

Returns a raw pointer to the MDBX environment.

pub fn id(&self) -> u64

Returns the transaction id.

pub fn get<Key>(&self, dbi: MDBX_dbi, key: &[u8]) -> Result<Option<Key>>
where Key: TableObject,

Gets an item from a database.

This function retrieves the data associated with the given key in the database. If the database supports duplicate keys (DatabaseFlags::DUP_SORT) then the first data item for the key will be returned. Retrieval of other items requires the use of Cursor. If the item is not in the database, then None will be returned.

pub fn commit(self) -> Result<(bool, CommitLatency)>

Commits the transaction.

Any pending operations will be saved.

pub fn prime_for_permaopen(&self, db: Database)

pub fn commit_and_rebind_open_dbs( self ) -> Result<(bool, CommitLatency, Vec<Database>)>

Commits the transaction and returns table handles permanently open until dropped.

pub fn open_db(&self, name: Option<&str>) -> Result<Database>

Opens a handle to an MDBX database.

If name is None, then the returned handle will be for the default database.

If name is not None, then the returned handle will be for a named database. In this case the environment must be configured to allow named databases through EnvironmentBuilder::set_max_dbs().

The returned database handle may be shared among any transaction in the environment.

The database name may not contain the null character.

pub fn db_flags(&self, db: &Database) -> Result<DatabaseFlags>

Gets the option flags for the given database in the transaction.

pub fn db_stat(&self, db: &Database) -> Result<Stat>

Retrieves database statistics.

pub fn db_stat_with_dbi(&self, dbi: MDBX_dbi) -> Result<Stat>

Retrieves database statistics by the given dbi.

pub fn cursor(&self, db: &Database) -> Result<Cursor<K>>

Open a new cursor on the given database.

pub fn cursor_with_dbi(&self, dbi: MDBX_dbi) -> Result<Cursor<K>>

Open a new cursor on the given dbi.

pub fn disable_timeout(&self)

Disables a timeout for this read transaction.

impl Transaction<RW>

pub fn create_db( &self, name: Option<&str>, flags: DatabaseFlags ) -> Result<Database>

Opens a handle to an MDBX database, creating the database if necessary.

If the database is already created, the given option flags will be added to it.

If name is None, then the returned handle will be for the default database.

If name is not None, then the returned handle will be for a named database. In this case the environment must be configured to allow named databases through EnvironmentBuilder::set_max_dbs().

This function will fail with Error::BadRslot if called by a thread with an open transaction.

pub fn put( &self, dbi: MDBX_dbi, key: impl AsRef<[u8]>, data: impl AsRef<[u8]>, flags: WriteFlags ) -> Result<()>

Stores an item into a database.

This function stores key/data pairs in the database. The default behavior is to enter the new key/data pair, replacing any previously existing key if duplicates are disallowed, or adding a duplicate data item if duplicates are allowed (DatabaseFlags::DUP_SORT).

pub fn reserve( &self, db: &Database, key: impl AsRef<[u8]>, len: usize, flags: WriteFlags ) -> Result<&mut [u8]>

Returns a buffer which can be used to write a value into the item at the given key and with the given length. The buffer must be completely filled by the caller.

pub fn del( &self, dbi: MDBX_dbi, key: impl AsRef<[u8]>, data: Option<&[u8]> ) -> Result<bool>

Delete items from a database. This function removes key/data pairs from the database.

The data parameter is NOT ignored regardless the database does support sorted duplicate data items or not. If the data parameter is Some only the matching data item will be deleted. Otherwise, if data parameter is None, any/all value(s) for specified key will be deleted.

Returns true if the key/value pair was present.

pub fn clear_db(&self, dbi: MDBX_dbi) -> Result<()>

Empties the given database. All items will be removed.

pub unsafe fn drop_db(&self, db: Database) -> Result<()>

Drops the database from the environment.

Safety

Caller must close ALL other Database and Cursor instances pointing to the same dbi BEFORE calling this function.

impl Transaction<RO>

pub unsafe fn close_db(&self, db: Database) -> Result<()>

Closes the database handle.

Safety

Caller must close ALL other Database and Cursor instances pointing to the same dbi BEFORE calling this function.

impl Transaction<RW>

pub fn begin_nested_txn(&mut self) -> Result<Transaction<RW>>

Begins a new nested transaction inside of this transaction.